fix: file-size + ghost-guard CI failures
- Extract build_default_provider to ci/mod.rs (watch.rs 787→749 LOC)
- Condense verbose comments to stay under 750 ceiling
- Add ci.notification_only to CONSUMED ghost guard with consumer cite
- Preserve all production + test behavior


Agend-Agent: claude-aef7c0
Agend-Task: t-20260718081644248223-72215-9
Agend-Branch: fix/2812-notification-only-watch
Agend-Issued-At: 2026-07-18T10:00:50.564800+00:00
